MOD bulk commit to setup Signed-off-by: Zdenek Styblik <stybla@turnovfree.net>
This is a bulk commit of modifications made to setup scripts. It brings new
features, new bugs and changes to the scripts resp. some shuffling around. It
also replaces some % echo; calls with % printf;.
Scripts utilize awk more now, because awk seems to be better fit to do the job
than shell.