FIX setup/SeTlilo - work-around for VirtIO disks Signed-off-by: Zdenek Styblik <stybla@turnovfree.net>
Commit fixes work-around for VirtIO disk. Prior to modification, all VirtIO
disks were added to '/etc/lilo.conf'. However, it turns out this is wrong and it
confuses LILO. No, it is just wrong. Result is wrong 'root=DEV' is specified and
kernel panics.
So from now on, params only for BOOTDEV should be added.